refactor: controller 수정
